requirements.txt
mss
numpy
Pillow
pywin32
